Data_lake_flask_app/flaskcode/requirements.txt
numpy
scipy
opencv-python
pandas
matplotlib
seaborn
imutils
imageio
qtpy
psutil
Flask
flask-wtf
flask-restful
gunicorn
pymongo
pymysql
sqlalchemy
requests
pyyaml
ruamel.yaml
notebook
jupyterlab